simplify boolean types
