+++ src/Type/BooleanType.php
@@ -2,7 +2,28 @@
 
 namespace PHPStan\Type;
 
-interface BooleanType extends Type
+use PHPStan\Type\Traits\NonCallableTypeTrait;
+use PHPStan\Type\Traits\NonIterableTypeTrait;
+use PHPStan\Type\Traits\NonObjectTypeTrait;
+use PHPStan\Type\Traits\NonOffsetAccessibleTypeTrait;
+
+class BooleanType implements Type
 {
+
+	use JustNullableTypeTrait;
+	use NonCallableTypeTrait;
+	use NonIterableTypeTrait;
+	use NonObjectTypeTrait;
+	use NonOffsetAccessibleTypeTrait;
+
+	public function describe(): string
+	{
+		return 'bool';
+	}
+
+	public static function __set_state(array $properties): Type
+	{
+		return new static();
+	}
 
 }
