--- src/Type/TrueOrFalseBooleanType.php
@@ -1,68 +1,0 @@
-<?php declare(strict_types = 1);
-
-namespace PHPStan\Type;
-
-use PHPStan\TrinaryLogic;
-use PHPStan\Type\Traits\NonCallableTypeTrait;
-use PHPStan\Type\Traits\NonIterableTypeTrait;
-use PHPStan\Type\Traits\NonObjectTypeTrait;
-use PHPStan\Type\Traits\NonOffsetAccessibleTypeTrait;
-
-class TrueOrFalseBooleanType implements BooleanType
-{
-
-	use NonCallableTypeTrait;
-	use NonIterableTypeTrait;
-	use NonObjectTypeTrait;
-	use NonOffsetAccessibleTypeTrait;
-
-	public function describe(): string
-	{
-		return 'bool';
-	}
-
-	/**
-	 * @return string[]
-	 */
-	public function getReferencedClasses(): array
-	{
-		return [];
-	}
-
-	public function accepts(Type $type): bool
-	{
-		if ($type instanceof BooleanType) {
-			return true;
-		}
-
-		if ($type instanceof CompoundType) {
-			return CompoundTypeHelper::accepts($type, $this);
-		}
-
-		return false;
-	}
-
-	public function isSuperTypeOf(Type $type): TrinaryLogic
-	{
-		if ($type instanceof BooleanType) {
-			return TrinaryLogic::createYes();
-		}
-
-		if ($type instanceof CompoundType) {
-			return $type->isSubTypeOf($this);
-		}
-
-		return TrinaryLogic::createNo();
-	}
-
-	public function getOffsetValueType(): Type
-	{
-		return new NullType();
-	}
-
-	public static function __set_state(array $properties): Type
-	{
-		return new self();
-	}
-
-}
